+++ namestnikov.kirill/M1/main.cpp
@@ -7,55 +7,11 @@
 #include <algorithm>
 #include <chrono>
 #include <functional>
-
-
-struct Point{
-public:
-  Point(double x, double y):
-    x_(x),
-    y_(y)
-  {}
-  double getX() const
-  {
-    return x_;
-  }
-  double getY() const
-  {
-    return y_;
-  }
-private:
-  double x_;
-  double y_;
-};
-
-using data_t = std::vector< Point >;
-using iterator_t = data_t::iterator;
-
-bool checkIfInCircle(const Point & point, int radius)
-{
-  return (std::pow(point.getX(), 2) + std::pow(point.getY(), 2) <= std::pow(radius, 2)); 
-}
-
-void generatePoints(std::default_random_engine & gen, size_t count, int radius, data_t & data)
-{
-  std::uniform_real_distribution< double > distribution(-radius, radius);
-  for (size_t i = 0; i < count; ++i)
-  {
-    double x = distribution(gen);
-    double y = distribution(gen);
-    Point p(x, y);
-    data.push_back(p);
-  }
-}
-
-size_t countPoints(const data_t & data, iterator_t begin, iterator_t end, int radius)
-{
-  using namespace std::placeholders;
-  return std::count_if(begin, end, std::bind(checkIfInCircle, _1, radius));
-}
+#include "calculate_square.hpp"
 
 double getSquareLin(std::default_random_engine & gen, size_t numberOfTests, int radius)
 {
+  using namespace namestnikov;
   std::uniform_real_distribution< double > distribution(-radius, radius);
   int res = 0;
   for (size_t i = 0; i < numberOfTests; ++i)
@@ -63,8 +19,8 @@ double getSquareLin(std::default_random_engine & gen, size_t numberOfTests, int 
     double x = distribution(gen);
     double y = distribution(gen);
     //std::cout << x << " " << y;
-    Point p(x, y);
-    if (checkIfInCircle(p, radius))
+    Point point(x, y);
+    if (std::pow(point.getX(), 2) + std::pow(point.getY(), 2) <= std::pow(radius, 2))
     {
       ++res;
     }
@@ -72,37 +28,9 @@ double getSquareLin(std::default_random_engine & gen, size_t numberOfTests, int 
   return 4 * std::pow(radius, 2) * res / numberOfTests;
 }
 
-double getSquare(std::default_random_engine & gen, size_t numberOfTests, int radius, size_t threadsCount)
-{
-  data_t data;
-  data.reserve(numberOfTests);
-  generatePoints(gen, numberOfTests, radius, data);
-  size_t possibleThreads = std::thread::hardware_concurrency();
-  threadsCount = std::min(possibleThreads, threadsCount);
-  std::vector< std::future< size_t > > fts;
-  fts.reserve(threadsCount - 1);
-  std::vector< size_t > results(threadsCount, 0);
-  size_t perThread = numberOfTests / threadsCount;
-  size_t lastThread = perThread + numberOfTests % threadsCount;
-  auto begin = data.begin();
-  for (size_t i = 0; i < threadsCount - 1; ++i)
-  {
-    auto end = begin + perThread;
-    fts.emplace_back(std::async(countPoints, std::cref(data), begin, end, radius));
-    begin = end;
-  }
-  results.back() = countPoints(data, begin, begin + lastThread, radius);
-  std::transform(fts.begin(), fts.end(), results.begin(),
-   [](auto && ft)
-   {
-    return ft.get();
-   });
-   size_t res = std::accumulate(results.begin(), results.end(), 0);
-   return 4 * std::pow(radius, 2) * res / numberOfTests; 
-}
-
 int main()
 {
+  using namespace namestnikov;
   size_t seed = 5;
   int radius = 2;
   size_t numberOfTests = 100000000;
